backend/requirements.txt
fastapi==0.115.0
python-jose[cryptography]==3.3.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
google-generativeai==0.8.3
python-dotenv==1.0.1
pdfplumber==0.11.4
python-docx==1.1.2
python-multipart==0.0.12
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
